LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-in-setTimeout-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: undefined_variable

PASS window.onerror - runtime error in setTimeout 
PASS window.onerror - runtime error in setTimeout (column) 
